requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
langchain-openai
chromadb
ipykernel
jupyter
sqlite-vss
